// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
const defaultOptions = {};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type Comment = {
  __typename?: "Comment";
  createdAt?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  isCommentMine: Scalars["Boolean"];
  text: Scalars["String"];
  tweet?: Maybe<Tweet>;
  updatedAt?: Maybe<Scalars["String"]>;
  user?: Maybe<User>;
};

export type CreatorsEntity = {
  __typename?: "CreatorsEntity";
  address?: Maybe<Scalars["String"]>;
  share?: Maybe<Scalars["Int"]>;
};

export type CreatorsEntityInput = {
  address?: Maybe<Scalars["String"]>;
  share?: Maybe<Scalars["Int"]>;
};

export type File = {
  __typename?: "File";
  id: Scalars["ID"];
  tweet?: Maybe<Tweet>;
  url: Scalars["String"];
  user?: Maybe<User>;
};

export type FilesEntity = {
  __typename?: "FilesEntity";
  type?: Maybe<Scalars["String"]>;
  uri?: Maybe<Scalars["String"]>;
};

export type FilesEntityInput = {
  type?: Maybe<Scalars["String"]>;
  uri?: Maybe<Scalars["String"]>;
};

export type Gif = {
  __typename?: "Gif";
  fixedHeightUrl: Scalars["String"];
  id: Scalars["ID"];
  originalUrl: Scalars["String"];
  title: Scalars["String"];
  tweet?: Maybe<Tweet>;
};
// ...
export type Nft = {
  __typename?: "NFT";
  attributes?: Maybe<Array<Maybe<AttributesEntity>>>;
  collection?: Maybe<Collection>;
  description?: Maybe<Scalars["String"]>;
  externalUrl?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  image: Scalars["String"];
  name?: Maybe<Scalars["String"]>;
  properties?: Maybe<Properties>;
  publicKey: Scalars["String"];
  sellerFeeBasisPoints?: Maybe<Scalars["Int"]>;
  symbol?: Maybe<Scalars["String"]>;
};
// ...
export type OneSignal = {
  __typename?: "OneSignal";
  createdAt?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  oneSignalId?: Maybe<Scalars["String"]>;
  updatedAt?: Maybe<Scalars["String"]>;
  user?: Maybe<User>;
};
// ...
export type Query = {
  __typename?: "Query";
  feed: Array<Tweet>;
  healthCheck: Scalars["String"];
  me: User;
  mentions: Array<Tweet>;
  oneSignal: OneSignal;
  profile: User;
  profileById: User;
  searchByTag: Array<Tweet>;
  searchByTweet: Array<Tweet>;
  searchByUser: Array<User>;
  tweet: Tweet;
  userFollow: Array<User>;
  users: Array<User>;
};
// ...
export type Reaction = {
  __typename?: "Reaction";
  count: Scalars["Int"];
  emojiId: Scalars["String"];
  id: Scalars["ID"];
  isMine: Scalars["Boolean"];
  skin?: Maybe<Scalars["Int"]>;
  tweet: Tweet;
  user: User;
};

export type Retweet = {
  __typename?: "Retweet";
  createdAt: Scalars["String"];
  id: Scalars["ID"];
  tweet?: Maybe<Tweet>;
  updatedAt: Scalars["String"];
  user?: Maybe<User>;
};

export type Tip = {
  __typename?: "Tip";
  createdAt?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  tipAmount: Scalars["String"];
  tweetId: Scalars["String"];
  updatedAt?: Maybe<Scalars["String"]>;
  userId: Scalars["String"];
};

export type TipInput = {
  tipAmount: Scalars["String"];
  tweetId: Scalars["String"];
  userId: Scalars["String"];
};

export type Tweet = {
  __typename?: "Tweet";
  childTweets?: Maybe<Array<Maybe<Tweet>>>;
  comments: Array<Comment>;
  commentsCount: Scalars["Int"];
  createdAt?: Maybe<Scalars["String"]>;
  files: Array<File>;
  gif?: Maybe<Gif>;
  id: Scalars["ID"];
  isRetweet: Scalars["Boolean"];
  isTweetMine: Scalars["Boolean"];
  mentions?: Maybe<Array<Scalars["String"]>>;
  nft?: Maybe<Nft>;
  parentTweet?: Maybe<Tweet>;
  reactions: Array<Reaction>;
  retweets: Array<Retweet>;
  retweetsCount: Scalars["Int"];
  tags: Array<Scalars["String"]>;
  text: Scalars["String"];
  tipsCount?: Maybe<Scalars["String"]>;
  updatedAt?: Maybe<Scalars["String"]>;
  user?: Maybe<User>;
};

export type User = {
  __typename?: "User";
  avatar: Scalars["String"];
  bio?: Maybe<Scalars["String"]>;
  comments: Array<Comment>;
  consumerName?: Maybe<Scalars["String"]>;
  coverPhoto?: Maybe<Scalars["String"]>;
  createdAt?: Maybe<Scalars["String"]>;
  dob?: Maybe<Scalars["String"]>;
  email?: Maybe<Scalars["String"]>;
  files: Array<File>;
  followers: Array<User>;
  followersCount: Scalars["Int"];
  following: Array<User>;
  followingCount: Scalars["Int"];
  handle: Scalars["String"];
  id: Scalars["ID"];
  isFollowing: Scalars["Boolean"];
  isSelf: Scalars["Boolean"];
  location?: Maybe<Scalars["String"]>;
  newMentionsCount: Scalars["Int"];
  nonce: Scalars["String"];
  publicAddress: Scalars["String"];
  reactions: Array<Reaction>;
  retweets: Array<Retweet>;
  tweets: Array<Tweet>;
  tweetsCount: Scalars["Int"];
  updatedAt?: Maybe<Scalars["String"]>;
  website?: Maybe<Scalars["String"]>;
};
// ...
export type ProfileQuery = {
  __typename?: "Query";
  profile: {
    __typename?: "User";
    id: string;
    publicAddress: string;
    handle: string;
    consumerName?: string | null | undefined;
    avatar: string;
    coverPhoto?: string | null | undefined;
    dob?: string | null | undefined;
    location?: string | null | undefined;
    website?: string | null | undefined;
    isSelf: boolean;
    isFollowing: boolean;
    followersCount: number;
    followingCount: number;
    tweetsCount: number;
    newMentionsCount: number;
    bio?: string | null | undefined;
    createdAt?: string | null | undefined;
    tweets: Array<{
      __typename?: "Tweet";
      id: string;
      text: string;
      tags: Array<string>;
      isTweetMine: boolean;
      commentsCount: number;
      retweetsCount: number;
      isRetweet: boolean;
      tipsCount?: string | null | undefined;
      createdAt?: string | null | undefined;
      user?:
        | {
            __typename?: "User";
            id: string;
            consumerName?: string | null | undefined;
            publicAddress: string;
            handle: string;
            avatar: string;
          }
        | null
        | undefined;
      files: Array<{ __typename?: "File"; id: string; url: string }>;
      gif?:
        | {
            __typename?: "Gif";
            id: string;
            title: string;
            fixedHeightUrl: string;
            originalUrl: string;
          }
        | null
        | undefined;
      reactions: Array<{
        __typename?: "Reaction";
        id: string;
        emojiId: string;
        skin?: number | null | undefined;
        isMine: boolean;
        count: number;
      }>;
    }>;
  };
};
// ...
export const ProfileDocument = gql`
  query profile($handle: String!) {
    profile(handle: $handle) {
      id
      publicAddress
      handle
      consumerName
      consumerName
      avatar
      coverPhoto
      dob
      location
      website
      isSelf
      isFollowing
      followersCount
      followingCount
      tweetsCount
      newMentionsCount
      bio
      tweets {
        id
        text
        tags
        isTweetMine
        user {
          id
          consumerName
          publicAddress
          handle
          avatar
        }
        files {
          id
          url
        }
        gif {
          id
          title
          fixedHeightUrl
          originalUrl
        }
        reactions {
          id
          emojiId
          skin
          isMine
          count
        }
        commentsCount
        retweetsCount
        isRetweet
        tipsCount
        createdAt
      }
      createdAt
    }
  }
`;